# openapi.yaml
# Regenerate the full openapi.yaml with corrected structure for optional parameters as query params
openapi: 3.0.0
info:
  title: Saju Backend API
  description: API for user signup, member fortune calculations, and saju palja analysis.
  version: 1.0.0

servers:
  - url: http://localhost:3000

paths:
  /users/signup:
    post:
      summary: Sign up a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, type, nickname, gender, birthday, birthdayType]
              properties:
                email:
                  type: string
                password:
                  type: string
                type:
                  type: string
                  enum: [USER, MEMBER]
                nickname:
                  type: string
                gender:
                  type: string
                  enum: [MALE, FEMALE]
                birthday:
                  type: string
                  example: "1991-02-06"
                birthdayType:
                  type: string
                  enum: [SOLAR, LUNAR]
                time:
                  type: string
                  example: "1200"
      responses:
        '200':
          description: User created

  /users/signin:
    post:
      summary: Sign in and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Signed in

  /users/me:
    get:
      summary: Get current user and their members
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User data with member info

  /manse/calculate:
    post:
      summary: Direct saju calculation without saving
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [birthday, gender, birthdayType]
              properties:
                birthday:
                  type: string
                  example: "1991-02-06"
                time:
                  type: string
                  example: "1200"
                gender:
                  type: string
                  enum: [MALE, FEMALE]
                birthdayType:
                  type: string
                  enum: [SOLAR, LUNAR]
      responses:
        '200':
          description: Saju and fortune result

  /manse/members/{memberId}/fortune:
    get:
      summary: Get saju + 운세 for saved member
      parameters:
        - in: path
          name: memberId
          required: true
          schema:
            type: string
        - in: query
          name: bigNum
          required: false
          schema:
            type: integer
        - in: query
          name: smallNum
          required: false
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Fortune breakdown

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT